Build binaries

+++ ci/build/build_graphdriver_plugins.sh
@@ -1,15 +1,23 @@
 #!/bin/bash
+if [ x"$GOPATH" == x"" ]; then
+	echo "GOPATH is not set!"
+	exit -1
+fi
+
 PLUGINS_REPO="github.com/atlantic777/docker_graphdriver_plugins"
 PLUGINS_REPO_PATH="$GOPATH/src/$PLUGINS_REPO"
 
-go get -v "$PLUGINS_REPO/..."
+mkdir -p "$PLUGINS_REPO_PATH" > /dev/null
+cp -r plugins/* "$PLUGINS_REPO_PATH"
 
-mkdir -p binaries
+mkdir -p binaries > /dev/null
 rm -rf binaries/*
 
-pushd binaries
+go get -v $PLUGINS_REPO/...
 
-for i in `ls $PLUGINS_REPO_PATH | grep "^[0-9]"`
+pushd binaries > /dev/null
+
+for i in `ls $PLUGINS_REPO_PATH | grep "cvmfs"`
 do
    echo "Build: $i"
    go build -v "$PLUGINS_REPO/$i"
